requirements.txt
pypng>=0.20220715.0
pyyaml>=6.0